Move arch specific code in ostd to the arch directory
