+++ ostd/src/arch/riscv/bus/mmio.rs
@@ -1,0 +1,3 @@
+// SPDX-License-Identifier: MPL-2.0
+
+pub(crate) fn init() {}
