+++ ostd/src/arch/riscv/bus/mod.rs
@@ -1,0 +1,3 @@
+// SPDX-License-Identifier: MPL-2.0
+
+pub(crate) mod mmio;
