+++ ostd/src/arch/x86/bus/mod.rs
@@ -1,0 +1,3 @@
+// SPDX-License-Identifier: MPL-2.0
+
+pub(crate) mod mmio;
