+++ ostd/src/arch/x86/mm/mod.rs
@@ -20,6 +20,7 @@ use crate::{
 
 mod util;
 
+pub(crate) const KERNEL_CODE_BASE_VADDR_PREFIX: usize = 0xffff_ffff_8000_0000;
 pub(crate) const NR_ENTRIES_PER_PAGE: usize = 512;
 
 #[derive(Clone, Debug, Default)]
