+++ ostd/src/bus/pci/mod.rs
@@ -58,18 +58,15 @@ mod device_info;
 
 pub use device_info::{PciDeviceId, PciDeviceLocation};
 
-use self::{bus::PciBus, common_device::PciCommonDevice};
-use crate::sync::Mutex;
+use crate::{
+    arch,
+    bus::pci::{bus::PciBus, common_device::PciCommonDevice},
+    sync::Mutex,
+};
 
 /// PCI bus instance
 pub static PCI_BUS: Mutex<PciBus> = Mutex::new(PciBus::new());
 
 pub(crate) fn init() {
-    let mut lock = PCI_BUS.lock();
-    for location in PciDeviceLocation::all() {
-        let Some(device) = PciCommonDevice::new(location) else {
-            continue;
-        };
-        lock.register_common_device(device);
-    }
+    arch::pci::init();
 }
