+++ ostd/src/cpu/mod.rs
@@ -4,17 +4,10 @@
 
 pub mod local;
 
-cfg_if::cfg_if! {
-    if #[cfg(target_arch = "x86_64")] {
-        pub use crate::arch::x86::cpu::*;
-    } else if #[cfg(target_arch = "riscv64")] {
-        pub use crate::arch::riscv::cpu::*;
-    }
-}
-
 use bitvec::prelude::BitVec;
 use spin::Once;
 
+pub use crate::arch::cpu::*;
 use crate::{
     arch::boot::smp::get_num_processors, cpu_local_cell, task::DisabledPreemptGuard,
     trap::DisabledLocalIrqGuard,
